finish mybatis

+++ src/main/java/com/example/comm/controller/PublishController.java
@@ -23,7 +23,7 @@ public class PublishController {
     private QuestionService questionService;
 
     @GetMapping("/publish/{id}")
-    public  String edit(@PathVariable(name = "id") Integer id,
+    public  String edit(@PathVariable(name = "id") Long id,
                         Model model){
         QuestionDto question=questionService.getById(id);
         model.addAttribute("title",question.getTitle());
